Set constants in ConstantsSetting class, delete magic words. Set cookie authorise system, update validation system.

+++ src/main/java/net/thumbtack/airline/controller/FlightController.java
@@ -1,10 +1,13 @@
 package net.thumbtack.airline.controller;
 
+import net.thumbtack.airline.ConstantsSetting;
 import net.thumbtack.airline.dto.request.FlightAddRequestDTO;
 import net.thumbtack.airline.dto.request.FlightGetParamsRequestDTO;
 import net.thumbtack.airline.dto.request.FlightUpdateRequestDTO;
 import net.thumbtack.airline.dto.response.FlightAddResponseDTO;
 import net.thumbtack.airline.dto.response.FlightUpdateResponseDTO;
+import net.thumbtack.airline.exception.SimpleException;
+import net.thumbtack.airline.service.CookieService;
 import net.thumbtack.airline.service.FlightService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
@@ -19,42 +22,68 @@ public class FlightController {
 
     private FlightService flightService;
 
+    private CookieService cookieService;
+
     @Autowired
     public void setFlightService(FlightService flightService) {
         this.flightService = flightService;
     }
 
+    @Autowired
+    public void setCookieService(CookieService cookieService) {
+        this.cookieService = cookieService;
+    }
+
     @PostMapping
-    public ResponseEntity<?> add(@RequestBody FlightAddRequestDTO request) {
+    public ResponseEntity<?> add(@RequestBody FlightAddRequestDTO request,
+                                 @CookieValue(value = "${cookie}", defaultValue = "") String uuid) {
+        if(uuid.isEmpty() || !cookieService.getUserCookie(uuid).getUserType().equals(ConstantsSetting.UserRoles.ADMIN_ROLE.toString())) {
+            throw new SimpleException(ConstantsSetting.ErrorsConstants.UNAUTHORISED_ERROR.toString(), "", "");
+        }
         FlightAddResponseDTO flightResponse =  flightService.add(request);
         return ResponseEntity.ok(flightResponse);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<?> update(@RequestBody FlightUpdateRequestDTO request) {
+    public ResponseEntity<?> update(@RequestBody FlightUpdateRequestDTO request,
+                                    @CookieValue(value = "${cookie}", defaultValue = "") String uuid) {
+        if(uuid.isEmpty() || !cookieService.getUserCookie(uuid).getUserType().equals(ConstantsSetting.UserRoles.ADMIN_ROLE.toString())) {
+            throw new SimpleException(ConstantsSetting.ErrorsConstants.UNAUTHORISED_ERROR.toString(), "", "");
+        }
         FlightUpdateResponseDTO flightResponse =  flightService.update(request);
         return ResponseEntity.ok(flightResponse);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> delete(@PathVariable("id") int id) {
+    public ResponseEntity<?> delete(@PathVariable("id") int id,
+                                    @CookieValue(value = "${cookie}", defaultValue = "") String uuid) {
+        if(uuid.isEmpty() || !cookieService.getUserCookie(uuid).getUserType().equals(ConstantsSetting.UserRoles.ADMIN_ROLE.toString())) {
+            throw new SimpleException(ConstantsSetting.ErrorsConstants.UNAUTHORISED_ERROR.toString(), "", "");
+        }
         flightService.delete(id);
         return ResponseEntity.ok().build();
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> get(@PathVariable("id") int id) {
+    public ResponseEntity<?> get(@PathVariable("id") int id,
+                                 @CookieValue(value = "${cookie}", defaultValue = "") String uuid) {
+        if(uuid.isEmpty() || !cookieService.getUserCookie(uuid).getUserType().equals(ConstantsSetting.UserRoles.ADMIN_ROLE.toString())) {
+            throw new SimpleException(ConstantsSetting.ErrorsConstants.UNAUTHORISED_ERROR.toString(), "", "");
+        }
         FlightAddResponseDTO flightResponse =  flightService.get(id);
         return ResponseEntity.ok(flightResponse);
     }
 
     @PutMapping("/{id}/approve")
-    public ResponseEntity<?> approve(@PathVariable("id") int id) {
+    public ResponseEntity<?> approve(@PathVariable("id") int id,
+                                     @CookieValue(value = "${cookie}", defaultValue = "") String uuid) {
+        if(uuid.isEmpty() || !cookieService.getUserCookie(uuid).getUserType().equals(ConstantsSetting.UserRoles.ADMIN_ROLE.toString())) {
+            throw new SimpleException(ConstantsSetting.ErrorsConstants.UNAUTHORISED_ERROR.toString(), "", "");
+        }
         FlightAddResponseDTO flightResponse =  flightService.approve(id);
         return ResponseEntity.ok(flightResponse);
     }
 
-    //Admin and user privileges
     @GetMapping
     public ResponseEntity<?> flights(
             @RequestParam(required = false, value = "fromTown", defaultValue = "") String fromTown,
@@ -62,7 +91,11 @@ public class FlightController {
             @RequestParam(required = false, value = "flightName ", defaultValue = "") String flightName,
             @RequestParam(required = false, value = "planeName", defaultValue = "") String planeName,
             @RequestParam(required = false, value = "fromDate", defaultValue = "") String fromDate,
-            @RequestParam(required = false, value = "toDate", defaultValue = "") String toDate) {
+            @RequestParam(required = false, value = "toDate", defaultValue = "") String toDate,
+            @CookieValue(value = "${cookie}", defaultValue = "") String uuid) {
+        if(uuid.isEmpty()) {
+            throw new SimpleException(ConstantsSetting.ErrorsConstants.UNAUTHORISED_ERROR.toString(), "", "");
+        }
 
         //TODO check user role. If admin, return flight with approve and plane
         List<FlightAddResponseDTO> flightResponse =  flightService.get(new FlightGetParamsRequestDTO(
